requirements.txt
flask
gunicorn
openai
pinecone
psycopg2-binary
python-dotenv
spacy==3.5.2
thinc==8.1.7
numpy==1.23.5
